+++ src/obf_params.h
@@ -12,8 +12,8 @@ typedef struct {
     size_t c;         // number of "symbolic" inputs
     size_t gamma;     // number of outputs
     size_t ell;       // length of symbols
-    uint32_t **types; // (gamma x (c+1))-size array
-    uint32_t M;       // max type degree in circuit over all output wires
+    size_t **types; // (gamma x (c+1))-size array
+    size_t M;       // max type degree in circuit over all output wires
     input_chunker chunker;
     reverse_chunker rchunker;
     circuit *circ;
